+++ NEWS.md
@@ -2,6 +2,8 @@
 - The codeclass.retain option now allows runLimmaAnalysis() to be run using
   a CodeClass/CodeClasses specified by the user, instead of automatically 
   removing non-endogenous genes. See help(runLimmaAnalysis) for details.
+- processNanostringData() can now handle a vector of .rcc files, in addition
+  to the previous options for loading NanoString data.
 
 ## Changes in NanoTube 1.5.1
 - We're published in Bioinformatics! CITATION file updated.